third_party/WebKit/LayoutTests/fast/events/eventlisteneroptions/capture_query.html

<!DOCTYPE HTML>
<body/>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>

<script>
test(function(t) {
    var supportsCapture = false;
    var query_function = function(e) {};
    var query_options = {
        get capture() {
            supportsCapture = true;
            return false;
        },
        get dummy() {
            assert_unreached("dummy value getter invoked");
            return false;
        }
    };

    document.addEventListener('test_event', query_function, query_options);
    assert_equals(supportsCapture, internals.runtimeFlags.eventListenerOptionsEnabled);
    supportsCapture = false;
    document.removeEventListener('test_event', query_function, query_options);
    assert_equals(supportsCapture, internals.runtimeFlags.eventListenerOptionsEnabled);
    t.done();
}, "Supports Capture");
</script>
